auth-service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
PyJWT==2.9.0
psycopg2-binary==2.9.9
gunicorn
Authlib>=1.3
requests>=2.31
